Refactor into proper methods
